* BufferContent.java (scan):  Add missing int->long cast. 	* emacs.java (point-at-eol):  Use buffer,getDot directly.
git-svn-id: svn://sourceware.org/svn/kawa/trunk@1532 5e0a886f-7f45-49c5-bc19-40643649e37f

+++ gnu/jemacs/buffer/BufferContent.java
@@ -476,7 +476,7 @@ implements javax.swing.text.AbstractDocument.Content
           }
         if (start > gapEnd)
           start -= gapEnd - gapStart;
-        return (count << 32) | start;
+        return ((long) count << 32) | start;
       }
     else
       {
